requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
attrs==23.2.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.19.0
google-api-python-client==2.127.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.63.0
httplib2==0.22.0
idna==3.7
multidict==6.0.5
proto-plus==1.23.0
protobuf==4.25.3
py-cord==2.5.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyNaCl==1.5.0
pyparsing==3.1.2
python-dotenv==1.0.1
requests==2.31.0
rsa==4.9
smmap==5.0.1
uritemplate==4.1.1
urllib3==2.2.1
yarl==1.9.4
